Extend Crypto.Algorand.Signature and split into modules
Problem:
1. Some signature types are defined outside of the Signature module,
which seems more appropriate for those types.
2. If we move them there, it becomes quite big.

Solution:
1. Move `SignatureType`, `MultiSignature` and `LogicSignature` into
`Signature.hs`.
2. Create Crypto.Algorand.{Key,Util} modules and extract some
definitions there.
